Add build configurations, use them for C/C++
To get a good experience when editing C/C++ (get accurate diagnostics,
locate includes, etc), one needs to inform the language server (e.g.
clangd) where the build resides.  The language server will then read the
compile_commands.json file from that directory, indicating, for each
file, how it has been compiled.  It is possible and common to have
multiple builds using the same source tree.

We therefore need a way in Theia to let the user select which build will
be used by the language server to get the compilation commands for the
files being edited.  Instead of doing something very C/C++ specific, I
added a build configurations service, that reads a list of build
configurations from the file .theia/builds.json, if it is present in the
workspace.  The user can then make a build active using the quick open
menu.

The C/C++ extension listens for build configuration changes and informs
clangd of the change.  The name of the active configuration is saved to
the local storage and restored when the frontend starts (if a config
with that name still exists).

The build configuration concept can be re-used for other languages if
necessary, and will be useful when we actually add the support to build
C/C++ code from Theia.

The old way of specifying the compile_commands.json path (through an env
var) is removed in favor of this new method.

Not part of this patch, but on the roadmap are:

- status bar icon to show the active build configuration
- watch .theia/builds.json to update the list of build configurations at
  runtime (currently, it is only read at startup)
- validate the builds.json file with a schema

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>