file-search: Allow searching in hidden files
When in the mode to search even .gitignored files, I think it would also
make sense to let the user search for hidden files.  For example, if you
want to open ".travis.yml".  This is easily achieved by passing a second
-u to ripgrep.  If we passed a third -u, it would also search through
binary files, I don't think we want to go there (yet).

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>